Modify 'call to action' page section
Implemented CSS styling for the call to action section on the landing
page.
Changed the div grouping within the 'cta' div to make the styling
easier.
Defined the 'Sign up' button styling for the hero section of the page
at the same time as both instances of the sign up button are the same
aside from border color.