Update README and adjust information centering
Previous version failed to include `justify-content: center;` for the
information section on the landing page. This resulted in the image
cards being left-aligned when resizing the page in a browser. Added
the correct alignment to rectify this issue.

Updated the README document to include a table of contents, a link to
the published page, and a retrospective section outlining some thoughts
for future improvement to the project.